requirements.txt
arrow==0.10.0
bleach==2.0.0
decorator==4.0.11
Django==1.11
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-extensions==1.7.8
djangorestframework==3.3.3
entrypoints==0.2.2
html5lib==0.999999999
ipykernel==4.6.1
ipython==6.0.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-client==5.0.1
jupyter-core==4.3.0
MarkupSafe==1.0
mistune==0.7.4
nbconvert==5.1.1
nbformat==4.3.0
notebook==5.0.0
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
python-dateutil==2.6.0
pytz==2017.2
pyzmq==16.0.2
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.3
terminado==0.6
testpath==0.3
tornado==4.5.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==2.0.0
